backend/requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
supabase==2.6.0
httpx==0.27.2
python-dotenv==1.0.1
rich==13.7.1
openai==1.51.0

